requirements.txt
aiohttp==3.5.4
async-timeout==3.0.1
attrs==19.3.0
chardet==3.0.4
discord.py==1.2.5
idna==2.8
multidict==4.7.4
pymongo==3.10.1
PyYAML==5.3
websockets==6.0
yarl==1.4.2
